requirements.txt
pandas==1.5.3
pandasai==0.8.1
python-dotenv==1.0.0
streamlit==1.25.0
